+++ src/api/chia/types/blockchain_format/serialized_program.ts
@@ -1,3 +1,0 @@
-import {str} from "../_python_types_";
-
-export type SerializedProgram = str;