+++ src/api/chia_rs/chia-bls/public_key.ts
@@ -1,0 +1,3 @@
+import {G1Element} from "../../chia/types/_python_types_";
+
+export type PublicKey = G1Element;