+++ src/api/chia_rs/chia-protocol/classgroup.ts
@@ -1,0 +1,5 @@
+import {bytes100} from "../../chia/types/blockchain_format/sized_bytes";
+
+export type ClassgroupElement = {
+  data: bytes100; // bytes100
+};