+++ src/api/chia_rs/chia-protocol/program.ts
@@ -1,0 +1,3 @@
+import {bytes} from "../../chia/types/_python_types_";
+
+export type Program = bytes;